refactor(cli): migrate argument handling to Cobra library
