+++ go.mod
@@ -1,3 +1,9 @@
 module github.com/gifflet/git-review
 
 go 1.20
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/cobra v1.9.1
+	github.com/spf13/pflag v1.0.6 // indirect
+)
